Move BaseArticle into bases.py
