+++ articles/models.py
@@ -1,75 +1,14 @@
 from django.conf import settings
-from django.db import models
-from django.db.models import Q
-from django.utils.translation import ugettext_lazy as _
 from django.core.urlresolvers import get_callable
-from django.conf.urls import patterns, url
-from django.utils.encoding import python_2_unicode_compatible
 
 try:
     from feincms.admin.item_editor import ItemEditor
 except ImportError:
     from feincms.admin.editor import ItemEditor
 
-from feincms.content.application import models as app_models
-from feincms.models import Base
-from feincms.module.mixins import ContentModelMixin
-from feincms.utils.managers import ActiveAwareContentManagerMixin
 from feincms.extensions import ExtensionModelAdmin
 
-
-class ArticleManager(ActiveAwareContentManagerMixin, models.Manager):
-    active_filters = {'simple-active': Q(active=True)}
-
-
-@python_2_unicode_compatible
-class BaseArticle(ContentModelMixin, Base):
-    active = models.BooleanField(_('active'), default=True)
-
-    title = models.CharField(_('title'), max_length=255)
-    slug = models.SlugField(_('slug'), max_length=255, help_text=_('This will be automatically generated from the name'), unique=True, editable=True)
-
-    class Meta:
-        ordering = ['title']
-        unique_together = []
-        verbose_name = _('article')
-        verbose_name_plural = _('articles')
-        abstract = True
-
-    objects = ArticleManager()
-
-    @classmethod
-    def get_urlpatterns(cls):
-        import views
-        return patterns('',
-            url(r'^$', views.ArticleList.as_view(), name='article_index'),
-            url(r'^(?P<slug>[a-z0-9_-]+)/$', views.ArticleDetail.as_view(), name='article_detail'),
-        )
-
-    @classmethod
-    def remove_field(cls, f_name):
-        """Remove a field. Effectively inverse of contribute_to_class"""
-        # Removes the field form local fields list
-        cls._meta.local_fields = [f for f in cls._meta.local_fields if f.name != f_name]
-
-        # Removes the field setter if exists
-        if hasattr(cls, f_name):
-            delattr(cls, f_name)
-
-    @classmethod
-    def get_urls(cls):
-        return cls.get_urlpatterns()
-
-    def __str__(self):
-        return self.title
-
-    @app_models.permalink
-    def get_absolute_url(self):
-        return ('article_detail', 'articles.urls', (), {'slug': self.slug})
-
-    @property
-    def is_active(self):
-        return self.__class__.objects.active().filter(pk=self.pk).count() > 0
+from articles.bases import BaseArticle
 
 
 class Article(BaseArticle):
